modularized the application to cope up with the size, created the products and shared module, getting Forms & Common Module from the shared module
