+++ src/app/products/models/product.interface.ts
