+++ src/app/products/products.module.ts
@@ -1,0 +1,27 @@
+import { NgModule } from "@angular/core";
+import { HttpClientModule } from "@angular/common/http";
+
+import { ProductService } from "./services/product.service";
+import { ProductListComponent } from "./products-list/product-list.component";
+import { SharedModule } from "../shared/shared.module";
+
+@NgModule({
+    imports: [
+        HttpClientModule,
+        SharedModule
+    ],
+    declarations: [
+        ProductListComponent
+    ],
+    providers: [
+        ProductService
+
+    ],
+    exports: [
+        ProductListComponent
+    ]
+
+})
+export class ProdusctsModule {
+
+}
