+++ src/app/products/services/product.service.ts
@@ -5,7 +5,7 @@ import 'rxjs/add/operator/do';
 import 'rxjs/add/operator/catch';
 import 'rxjs/add/observable/throw';
 
-import { IProduct } from "../product.interface";
+import { IProduct } from "../models/product.interface";
 
 
 @Injectable()
