+++ src/app/shared/pipes/convert-to-spaces.pipe.ts
