+++ src/app/shared/star/star-component.css
