+++ src/app/shared/star/star-component.ts
